# Skeleton system optimization