--- a/UMI3D-SDK/Assets/UMI3D SDK/ClientDevelopmentKit/Collaboration/Runtime/UserCapture/CollaborationSkeletonsManager.cs
+++ b/UMI3D-SDK/Assets/UMI3D SDK/ClientDevelopmentKit/Collaboration/Runtime/UserCapture/CollaborationSkeletonsManager.cs
@@ -142,11 +142,12 @@ namespace umi3d.cdk.collaboration.userCapture
         #region LifeCycle
 
         private bool canClearSkeletons = false;
+        private bool canUpdateSkeletons = false;
 
         private void Init()
         {
             collaborativeEnvironmentManagementService.OnUpdateJoinnedUserList += () => UpdateSkeletons(collaborativeEnvironmentManagementService.UserList);
-            collaborativeLoaderService.onEnvironmentLoaded.AddListener(() => { InitSkeletons(); if (ShouldSendTracking) SendTrackingLoop(); canClearSkeletons = true; });
+            collaborativeLoaderService.onEnvironmentLoaded.AddListener(() => { InitSkeletons(); if (ShouldSendTracking) SendTrackingLoop(); canClearSkeletons = true; canUpdateSkeletons = true; });
             collaborationClientServerService.OnLeavingEnvironment.AddListener(Clear);
             collaborationClientServerService.OnRedirection.AddListener(Clear);
         }
@@ -299,7 +300,7 @@ namespace umi3d.cdk.collaboration.userCapture
 
         public virtual void UpdateSkeleton(IEnumerable<UserTrackingFrameDto> frames)
         {
-            if (Application.isBatchMode)
+            if (Application.isBatchMode || !canUpdateSkeletons)
                 return;
 
             if (frames is null)