--- a/UMI3D-SDK/Assets/UMI3D SDK/ClientDevelopmentKit/Core/Runtime/Binding/Objects/AbstractSimpleBinding.cs
+++ b/UMI3D-SDK/Assets/UMI3D SDK/ClientDevelopmentKit/Core/Runtime/Binding/Objects/AbstractSimpleBinding.cs
@@ -29,7 +29,7 @@ namespace umi3d.cdk.binding
     {
         #region DTO access
 
-        protected AbstractSimpleBindingDataDto SimpleBindingData => data as AbstractSimpleBindingDataDto;
+        protected AbstractSimpleBindingDataDto SimpleBindingData { get; }
 
         /// <summary>
         /// See <see cref="AbstractSimpleBindingDataDto.syncPosition"/>.
@@ -80,6 +80,7 @@ namespace umi3d.cdk.binding
             OffSetRotation = SimpleBindingData.offSetRotation.Quaternion();
             OffSetScale = SimpleBindingData.offSetScale.Struct();
             AnchorPosition = SimpleBindingData.anchorPosition.Struct();
+            SimpleBindingData = (AbstractSimpleBindingDataDto)data;
         }
 
         /// <summary>