--- a/UMI3D-SDK/Assets/UMI3D SDK/ClientDevelopmentKit/Core/Runtime/Binding/Objects/MultiBinding.cs
+++ b/UMI3D-SDK/Assets/UMI3D SDK/ClientDevelopmentKit/Core/Runtime/Binding/Objects/MultiBinding.cs
@@ -31,7 +31,7 @@ namespace umi3d.cdk.binding
         /// </summary>
         protected readonly AbstractSimpleBinding[] orderedBindings;
 
-        public IReadOnlyList<AbstractSimpleBinding> Bindings => orderedBindings.ToList();
+        public IReadOnlyList<AbstractSimpleBinding> Bindings { get; }
 
         public MultiBinding(MultiBindingDataDto data, AbstractSimpleBinding[] bindings, Transform boundTransform, bool isOrdered = false) : base(boundTransform, data)
         {
@@ -45,6 +45,7 @@ namespace umi3d.cdk.binding
                                                .OrderByDescending(x => x.Priority)
                                                .ToArray();
             }
+            Bindings = orderedBindings.ToList();
         }
 
         /// <inheritdoc/>
@@ -80,7 +81,7 @@ namespace umi3d.cdk.binding
             if (boundTransform == null) // object destroyed before binding
                 return;
 
-            foreach (AbstractSimpleBinding binding in Bindings)
+            foreach (AbstractSimpleBinding binding in orderedBindings)
             {
                 if (binding is not null && binding.ResetWhenRemoved)
                 {