--- a/UMI3D-SDK/Assets/UMI3D SDK/ClientDevelopmentKit/Core/Runtime/Binding/Objects/NodeBinding.cs
+++ b/UMI3D-SDK/Assets/UMI3D SDK/ClientDevelopmentKit/Core/Runtime/Binding/Objects/NodeBinding.cs
@@ -30,9 +30,10 @@ namespace umi3d.cdk.binding
         public NodeBinding(NodeBindingDataDto dto, Transform boundTransform, UMI3DNodeInstance parentNode) : base(dto, boundTransform)
         {
             this.parentNode = parentNode;
+            NodeBindingDataDto = (NodeBindingDataDto)SimpleBindingData;
         }
 
-        protected NodeBindingDataDto NodeBindingDataDto => SimpleBindingData as NodeBindingDataDto;
+        protected NodeBindingDataDto NodeBindingDataDto { get; }
 
         #region DTO Access