--- a/UMI3D-SDK/Assets/UMI3D SDK/ClientDevelopmentKit/UserCapture/Runtime/Skeleton/Binding/Objects/BoneBinding.cs
+++ b/UMI3D-SDK/Assets/UMI3D SDK/ClientDevelopmentKit/UserCapture/Runtime/Skeleton/Binding/Objects/BoneBinding.cs
@@ -33,11 +33,12 @@ namespace umi3d.cdk.userCapture.binding
         public BoneBinding(BoneBindingDataDto dto, Transform boundTransform, ISkeleton skeleton) : base(dto, boundTransform)
         {
             this.skeleton = skeleton;
+            BoneBindingDataDto = (BoneBindingDataDto)SimpleBindingData;
         }
 
         #region DTO Access
 
-        protected BoneBindingDataDto BoneBindingDataDto => (BoneBindingDataDto)SimpleBindingData;
+        protected BoneBindingDataDto BoneBindingDataDto { get; }
 
         /// <summary>
         /// See <see cref="BoneBindingDataDto.userId"/>.